call controller action hello and render header page
