+++ config/web.php
@@ -47,8 +47,8 @@ $config = [
         ],
         'db' => $db,
         'urlManager' => [
-            'enablePrettyUrl' => true,
-            'showScriptName' => false
+            'enablePrettyUrl' => false,
+            'showScriptName' => true
         ]
     ],
     'modules' => [
