+++ controllers/PostCommentController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace app\controllers;
+
+use yii\web\Controller;
+
+class PostCommentController extends Controller{
+    public function actionIndex(){
+        return __METHOD__;
+    }
+}
