+++ controllers/TestController.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace app\controllers;
+
+
+use yii\web\Controller;
+
+class TestController extends Controller
+{
+    public function actionIndex(){
+        return '<h1>Hello World!</h1>';
+    }
+}
