+++ controllers/admin/MyController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace app\controllers\admin;
+
+use yii\web\Controller;
+
+class MyController extends Controller{
+    public function actionIndex(){
+        return __METHOD__;
+    }
+}
