+++ views/site/hello.php
@@ -7,4 +7,4 @@
  */
 ?>
 
-<h1>Welcome</h1>
+<h1>Welcom213e</h1>
